separando arquitetura do projeto
