+++ src/index.js
@@ -1,47 +1,16 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import Header from "./Header"
-import User from './User';
-import Content from './Content';
-import Menu from './Menu';
-
-let deferredPrompt;
-const addBtn = document.querySelector(".add-button");
-addBtn.style.display = "none";
-
-window.addEventListener("beforeinstallprompt", (e) => {
-  // Prevent Chrome 67 and earlier from automatically showing the prompt
-  e.preventDefault();
-  // Stash the event so it can be triggered later.
-  deferredPrompt = e;
-  // Update UI to notify the user they can add to home screen
-  addBtn.style.display = "block";
-
-  addBtn.addEventListener("click", (e) => {
-    // hide our user interface that shows our A2HS button
-    addBtn.style.display = "none";
-    // Show the prompt
-    deferredPrompt.prompt();
-    // Wait for the user to respond to the prompt
-    deferredPrompt.userChoice.then((choiceResult) => {
-      if (choiceResult.outcome === "accepted") {
-        console.log("User accepted the A2HS prompt");
-      } else {
-        console.log("User dismissed the A2HS prompt");
-      }
-      deferredPrompt = null;
-    });
-  });
-});
+import App from './App';
 
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
-  <React.StrictMode> 
-        <Header />
-        <User/>
-        <Menu />
-        <Content />
+  <React.StrictMode>
+    <App />
   </React.StrictMode>
 );
+
+// If you want to start measuring performance in your app, pass a functaion
+// to log results (for example: reportWebVitals(console.log))
+// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
